+++ src/transactions/types/htlc_refund.cpp
@@ -36,7 +36,7 @@ namespace transactions {
 // Lock Transaction Id - 32 Bytes:
 // - std::move(buffer, buffer + 32, refund->id.begin());
 // ---
-auto HtlcRefund::Deserialize(HtlcRefund *refund, const uint8_t *buffer) -> uint32_t {
+auto HtlcRefund::Deserialize(HtlcRefund *refund, const uint8_t *buffer) -> size_t {
     refund->id = {};
     std::move(buffer, buffer + HASH_32_LEN, refund->id.begin());    // 32 Bytes
 
@@ -58,7 +58,7 @@ auto HtlcRefund::Deserialize(HtlcRefund *refund, const uint8_t *buffer) -> uint3
 // - std::move(refund.id.begin(), refund.id.end(), buffer);
 //
 // ---
-auto HtlcRefund::Serialize(const HtlcRefund &refund, uint8_t *buffer) -> uint32_t {
+auto HtlcRefund::Serialize(const HtlcRefund &refund, uint8_t *buffer) -> size_t {
     std::move(refund.id.begin(), refund.id.end(), buffer);          // 32 Bytes
 
     return HASH_32_LEN;
